Added method to import TJDS matrix

+++ src/JDS.cpp
@@ -1,6 +1,6 @@
 #include "JDS.h"
 
-static std::vector<unsigned int> per_row_nonzeros(const std::vector<std::vector<float>>& mat)
+static std::vector<unsigned int> per_row_nonzeros(const std::vector<std::vector<double>> &mat)
 {
     // For each row, increment a counter for each non-zero value stored in the matrix and assign that row's nonzero counter in the output vector.
     std::vector<unsigned int> nonzeros(mat.size());
@@ -24,11 +24,11 @@ static std::vector<unsigned int> permute_descending(std::vector<unsigned int> no
 {
     // Use selection sort to set indices in nonzero-count-descending order.
     std::vector<unsigned int> permutation(nonzeros.size());
-    for (unsigned int i = 0; i < permutation.size(); i++) 
+    for (unsigned int i = 0; i < permutation.size(); i++)
     {
         permutation[i] = i;
     }
-    
+
     for (unsigned int i = 0; i < permutation.size(); i++)
     {
         unsigned int max_idx = i;
@@ -39,7 +39,7 @@ static std::vector<unsigned int> permute_descending(std::vector<unsigned int> no
                 max_idx = j;
             }
         }
-        
+
         if (max_idx != i)
         {
             unsigned int temp = permutation[i];
@@ -47,52 +47,55 @@ static std::vector<unsigned int> permute_descending(std::vector<unsigned int> no
             permutation[max_idx] = temp;
         }
     }
-    
+
     return permutation;
 }
 
-JDSMatrix to_jds(std::vector<std::vector<float>> mat)
+JDSMatrix to_jds(std::vector<std::vector<double>> mat)
 {
     struct JDSMatrix out;
-    out.jdiag = std::vector<float>();
+    out.jdiag = std::vector<double>();
     out.col_ind = std::vector<unsigned int>();
     out.perm = std::vector<unsigned int>();
     out.jd_ptr = std::vector<unsigned int>();
     out.num_rows = mat.size();
     out.num_cols = mat.empty() ? 0 : mat[0].size();
-    if (out.num_rows == 0 || out.num_cols == 0) return out;
-    
+    if (out.num_rows == 0 || out.num_cols == 0)
+        return out;
+
     std::vector<unsigned int> nonzeros = per_row_nonzeros(mat);
     std::vector<unsigned int> forward_perm = permute_descending(nonzeros);
-    
+
     // Store forward permutation as 1-based in out.perm
     for (unsigned int i = 0; i < forward_perm.size(); i++)
     {
-        out.perm.push_back(forward_perm[i] + 1);  // Convert to 1-based
+        out.perm.push_back(forward_perm[i] + 1); // Convert to 1-based
     }
-    
+
     unsigned int max_nonzeros = 0;
-    for (unsigned int i = 0; i < nonzeros.size(); i++) 
-        if (nonzeros[i] > max_nonzeros) max_nonzeros = nonzeros[i];
-    
+    for (unsigned int i = 0; i < nonzeros.size(); i++)
+        if (nonzeros[i] > max_nonzeros)
+            max_nonzeros = nonzeros[i];
+
     if (max_nonzeros == 0)
     {
         out.jd_ptr.push_back(1);
         return out;
     }
-    
-    std::vector<std::vector<std::pair<unsigned int, float>>> rows(out.num_rows);
+
+    std::vector<std::vector<std::pair<unsigned int, double>>> rows(out.num_rows);
     for (unsigned int i = 0; i < out.num_rows; i++)
     {
         unsigned int orig = forward_perm[i];
         rows[i].reserve(nonzeros[orig]);
         for (unsigned int j = 0; j < out.num_cols; j++)
         {
-            if (mat[orig][j] == 0.) continue;
+            if (mat[orig][j] == 0.)
+                continue;
             rows[i].push_back(std::make_pair(j, mat[orig][j]));
         }
     }
-    
+
     for (unsigned int k = 0; k < max_nonzeros; k++)
     {
         out.jd_ptr.push_back(out.jdiag.size() + 1);
@@ -105,32 +108,32 @@ JDSMatrix to_jds(std::vector<std::vector<float>> mat)
             }
         }
     }
-    
+
     return out;
 }
 
-std::vector<std::vector<float>> from_jds(struct JDSMatrix jds)
+std::vector<std::vector<double>> from_jds(struct JDSMatrix jds)
 {
-    std::vector<std::vector<float>> output(jds.num_rows, std::vector<float>(jds.num_cols, 0));
+    std::vector<std::vector<double>> output(jds.num_rows, std::vector<double>(jds.num_cols, 0));
     if (jds.num_rows == 0 || jds.num_cols == 0)
     {
         return output;
     }
-    
+
     unsigned int num_diags = jds.jd_ptr.size();
-    
+
     for (unsigned int i = 0; i < jds.num_rows; i++)
     {
-        unsigned int orig = jds.perm[i] - 1;  // Convert from 1-based to 0-based
+        unsigned int orig = jds.perm[i] - 1; // Convert from 1-based to 0-based
         for (unsigned int k = 0; k < num_diags; k++)
         {
             unsigned int start = jds.jd_ptr[k] - 1;
             unsigned int end;
-            if(k + 1 < jds.jd_ptr.size())
+            if (k + 1 < jds.jd_ptr.size())
             {
                 end = jds.jd_ptr[k + 1] - 1;
-            } 
-            else 
+            }
+            else
             {
                 end = jds.jdiag.size();
             }
@@ -143,29 +146,30 @@ std::vector<std::vector<float>> from_jds(struct JDSMatrix jds)
             }
         }
     }
-    
+
     return output;
 }
 
-std::vector<float> jds_matrix_vector_mult(struct JDSMatrix jds, std::vector<float> x)
+std::vector<double> jds_matrix_vector_mult(struct JDSMatrix jds, std::vector<double> x)
 {
-    std::vector<float> y(jds.num_rows, 0.0f);
-    if (jds.num_rows == 0 || jds.num_cols == 0) return y;
-    
+    std::vector<double> y(jds.num_rows, 0.0f);
+    if (jds.num_rows == 0 || jds.num_cols == 0)
+        return y;
+
     unsigned int num_diags = jds.jd_ptr.size();
-    
+
     for (unsigned int i = 0; i < jds.num_rows; i++)
     {
-        unsigned int orig = jds.perm[i] - 1;  // Convert from 1-based to 0-based
+        unsigned int orig = jds.perm[i] - 1; // Convert from 1-based to 0-based
         for (unsigned int k = 0; k < num_diags; k++)
         {
             unsigned int start = jds.jd_ptr[k] - 1;
             unsigned int end;
-            if(k + 1 < jds.jd_ptr.size())
+            if (k + 1 < jds.jd_ptr.size())
             {
                 end = jds.jd_ptr[k + 1] - 1;
-            } 
-            else 
+            }
+            else
             {
                 end = jds.jdiag.size();
             }
@@ -178,6 +182,6 @@ std::vector<float> jds_matrix_vector_mult(struct JDSMatrix jds, std::vector<floa
             }
         }
     }
-    
+
     return y;
 }
